pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fund-manager"
version = "0.1.0"
description = "Personal, self-hosted, agent-driven fund portfolio review and strategy assistant."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "Fund Manager Maintainers" }]
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "jinja2>=3.1.4",
  "psycopg2-binary>=2.9.11",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.22",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
data = [
  "akshare>=1.15.0",
  "pandas>=2.2.3",
]
mcp = [
  "mcp>=1.6.0",
]
dev = [
  "mypy>=1.13.0",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
]

[project.scripts]
fund-manager-api = "fund_manager.apps.api.main:run"
fund-manager-admin = "fund_manager.admin.cli:main"
fund-manager-mcp = "fund_manager.mcp.server:main"
fund-manager-scheduler = "fund_manager.scheduler.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.package-data]
fund_manager = [
  "agents/prompts/*.md",
  "reports/templates/*.j2",
  "apps/api/templates/*.html",
  "data/*.json",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["ASYNC", "B", "E", "F", "I", "N", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
show_error_codes = true
explicit_package_bases = true
mypy_path = "src"
packages = ["fund_manager"]
